src/services/backend/progress_tracker_microservice/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
asyncpg 